+++ e2etest/test_manager.go
@@ -12,6 +12,7 @@ import (
 	"go.uber.org/zap"
 	"os"
 	"path/filepath"
+	"slices"
 	"testing"
 	"time"
 
@@ -56,6 +57,7 @@ func defaultVigilanteConfig() *config.Config {
 type TestManager struct {
 	TestRpcClient   *rpcclient.Client
 	BitcoindHandler *BitcoindTestHandler
+	Electrs         *ElectrsTestHandler
 	BabylonClient   *bbnclient.Client
 	BTCClient       *btcclient.Client
 	Config          *config.Config
@@ -88,11 +90,16 @@ func StartManager(t *testing.T, numMatureOutputsInWallet uint32, epochInterval u
 	require.NoError(t, err)
 
 	btcHandler := NewBitcoindHandler(t, manager)
-	bitcoind := btcHandler.Start(t)
+	bitcoind, bitcoindPath := btcHandler.Start(t)
 	passphrase := "pass"
 	_ = btcHandler.CreateWallet("default", passphrase)
 
+	internalBtcRpc := fmt.Sprintf("%s:18443", bitcoind.Container.NetworkSettings.IPAddress)
+	electrsHandler := NewElectrsHandler(t, manager)
+	electrs := electrsHandler.Start(t, bitcoindPath, internalBtcRpc)
+
 	cfg := defaultVigilanteConfig()
+	cfg.BTCStakingTracker.IndexerAddr = fmt.Sprintf("http://localhost:%s", electrs.GetPort("3000/tcp"))
 
 	cfg.BTC.Endpoint = fmt.Sprintf("127.0.0.1:%s", bitcoind.GetPort("18443/tcp"))
 
@@ -107,7 +114,7 @@ func StartManager(t *testing.T, numMatureOutputsInWallet uint32, epochInterval u
 	}, nil)
 	require.NoError(t, err)
 
-	err = testRpcClient.WalletPassphrase(passphrase, 600)
+	err = testRpcClient.WalletPassphrase(passphrase, 60000)
 	require.NoError(t, err)
 
 	walletPrivKey, err := importPrivateKey(btcHandler)
@@ -161,6 +168,7 @@ func StartManager(t *testing.T, numMatureOutputsInWallet uint32, epochInterval u
 		TestRpcClient:   testRpcClient,
 		BabylonClient:   babylonClient,
 		BitcoindHandler: btcHandler,
+		Electrs:         electrsHandler,
 		BTCClient:       btcClient,
 		Config:          cfg,
 		WalletPrivKey:   walletPrivKey,
@@ -236,8 +244,11 @@ func (tm *TestManager) CatchUpBTCLightClient(t *testing.T) {
 		headers = append(headers, header)
 	}
 
-	_, err = tm.InsertBTCHeadersToBabylon(headers)
-	require.NoError(t, err)
+	for headersChunk := range slices.Chunk(headers, 1500) {
+		_, err := tm.InsertBTCHeadersToBabylon(headersChunk)
+		require.NoError(t, err)
+
+	}
 }
 
 func importPrivateKey(btcHandler *BitcoindTestHandler) (*btcec.PrivateKey, error) {